[benchmark projects] rename files
