+++ benchmark/projects/eval_test.go
@@ -1,4 +1,4 @@
-package project
+package projects
 
 import (
 	"testing"
