+++ benchmark/projects/native_test.go
@@ -1,4 +1,4 @@
-package project
+package projects
 
 import (
 	"github.com/onheap/eval_lab/benchmark"
